requirements.txt
# Python 3.8 Compatible Requirements
# Install with: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
pillow>=9.5.0
numpy>=1.20.0,<1.25.0
pandas>=1.3.0,<2.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
accelerate>=0.20.0
datasets>=2.10.0
scikit-learn>=1.0.0
tqdm>=4.65.0
requests>=2.28.0
opencv-python>=4.5.0
sentencepiece>=0.1.99
protobuf>=3.20.0
einops>=0.6.0
timm>=0.6.0
bitsandbytes>=0.41.0
scipy>=1.7.0

